WL#14710: InnoDB: Introducing REDUCED double write mode (Facebook Contribution)
Currently, there is a system variable --innodb-doublewrite to enable and
disable doublewrite buffers.  This worklog introduces two new values for this
system variable DETECT_ONLY and DETECT_AND_RECOVER.  The DETECT_AND_RECOVER
value is synonymous to the ON value of this variable.

In this detect_only mode, doublewrite buffer will only be used to detect a page
corruption and will not be used to recover from the corruption.  The double
write buffer will not contain the actual page contents.  This detect-only mode
helps to distinguish from known partial-write corruptions vs disk corruption
issues.

This patch is contributed by Facebook.  The mode was called REDUCED in the
patch contributed by Facebook.  We renamed the user visible name to be
DETECT_ONLY.

rb#26722 approved by Yasufumi Kinoshita <yasufumi.kinoshita@oracle.com>