Bug#33719861 FIXES NDB_RESTORE_APPLY_STATUS TEST
Fixes ndb_restore tool to generate ndb_apply_status table's
row with server_id=0 apprpriately.

Bug#32604161 introduced a feature in ndb_restore to allow
the users to optionally restore data of the system table,
ndb_apply_status. After all the rows were restored, the
row with server_id=0 was deleted to let the restore-epoch
option generate it appropriately. So, when ndb_restore is
invoked with both with-apply-status and restore-epoch
options, the generation of the row was supposed to happen
after the DELETE operation serially since these operations
were meant to be performed by a single thread. But, the
flag which makes sure the DELETE operation is performed by
a single thread wasn't initialised properly which in turn
resulted in more than one thread attempting the DELETE
operation and deleting the generated row by restore-epoch
option. So, the flag is set properly beforehand to handle
the described issue.

Change-Id: Ib3957e97aa927eb9aa5bf6653a5c4a06aeb03663