Bug #33789526 srv_error_monitor_thread shouldn't call functions which can deadlock
This patch fixes the main issue, by moving the srv_refresh_innodb_monitor_stats()
call from srv_error_monitor_thread() which shouldn't bother with it, to the
srv_monitor_thread() which really needs it. This "helping" use to have
some sense in the past when the srv_monitor_thread could be suspended,
while we still needed to keep refreshing the stats, but in todays trunk
the srv_monitor_thread never gets suspended.

Also, the patch simplifies the implementation based on observation that
current implementation in trunk doesn't really react to os_event_set
calls from various places in the way it was supposed to in the old
times. Originally, the intention was to wake up a suspended monitor
thread when we enable the monitoring. But, at least since 5.0 import
commit, we never really suspend the monitoring thread, and all this
os_event_set achieves is that it needlessly wakes up the monitoring
thread which will check the clock, notice that the time that passed is
smaller than 15 seconds, and will get back to sleep again.
So, we can remove all these os_event_set complication.
Also, there's only one thread which calls the os_event_reset.
Also, after the changes, we only ever call os_event_set on shutdown.
Combining these observations we can simplify the code further, by
changing the os_event_wait timeout to 15 seconds, and interpreting the
event as a signal to quit.

Renamed srv_last_monitor_time to srv_monitor_stats_refreshed_at.

When innodb needs to enable or disable the monitor, it should be done
independently to the user setting the sys var, and properly counting the
number of threads which wanted to enabled/disable it.
The neew srv_innodb_needs_monitoring atomic will keep track of it.
This change also fixes:
Bug #93878 innodb_status_output fails to restore to old value

Fixed a bug in shutdown, that we forgot to signal 4 threads:
- srv_threads.m_error_monitor
- srv_threads.m_monitor
- srv_threads.m_buf_dump
- srv_threads.m_buf_resize
This wasn't a big deal as some of them were noticing the shutdown
themselves, when waiting on event timed out, and others typically
wasn't even started at the moment the shutdown occurred.

Reviewed by: Marcin Babij <marcin.babij@oracle.com>
Change-Id: Ie483e6993aae54a069d0806255afbcd70e33b7e3
Change-Id: I3e28240030f7a167e0ede2a55a4f2ec7d79ab55b
Change-Id: Ia1a9cba1953cf1f6fd58430676968f93f7775756
Change-Id: I3d4602f34ecddc39d1d4a7e88399124ba465c3d2
Change-Id: Iee0af945bbaf107178690577b6aa6bed3c177ff6
Change-Id: I1356decfe54b2987eaad7b64dbd5b1214e5f230c
Change-Id: I554266fd1ce9ea1e43162f406981e8d31d506664

Change-Id: Ib8838647a756f297f7f9808eff8ab85f46d1b3ca