BUG#33830926: rpl_gtid.rpl_mts_spco_acl_commands_binlog fails on pb2
Symptom:

rpl_mts_spco_acl_commands_binlog and
rpl_mts_spco_acl_commands_nobinlog failed after WL#13469.

Analysis:

The tests replicates a CREATE USER statement, which is held before
committing due to replica-preserve-commit-order.  At that point, it
holds the metadata lock ACL CACHE in EXCLUSIVE mode.  While CREATE
USER is in this state, the test executes a number of checks, one of
which is a SHOW GLOBAL VARIABLES statement.

WL#13469 changed the execution of SHOW GLOBAL VARIABLES so that it
needs to take the metadata lock ACL CACHE (in SHARED mode).
Therefore, the execution of SHOW GLOBAL VARIABLES blocked the test.
Since the test could not proceed, it did not reach the point where it
unblocked the applier threads.  Eventually the applier threads timed
out, which made them stop with an error, and also abort the CREATE
USER statement.  Aborting CREATE USER also unblocked the SHOW GLOBAL
VARIABLES, which allowed the test to proceed.  But a little bit later,
the test failed since it did not expect an error in the replication
threads.

Fix:

Change the test to use SELECT @@global.variable instead of SHOW GLOBAL
VARIABLES LIKE 'variable'.

Change-Id: Id6772e2dabbcbd5627c846d5bb3c3cfe2a1279c8