Bug#33803541 Excessive CPU consumed by Dblqh::findTransaction()
When there are many operations belonging to still open transactions in
the hash list searched by Dblqh::findTransaction(), the hash buckets
becomes too filled. Thus there are too much CPU cycles consumed searcing
through the hash buckets.

The number of hash buckets are fixed to 4095, which is relatively
small considdered that we often configure the max number of operations
to much more than the default of 32k

This patch scales the size of the hash bucket relative to the
configured max number of operations, such that only a few items
should hash into the same bucket.

The same test case as reported in the bug report reduced its
elapsed run time to ~50%, and demonstrated a relative stable
insert rate: (Which decreased over time before this patch)

The select below select columns from a table just inserted into
in a single large transaction. Left column is the timestamp of
the insert with 100ms resolution, right column is number of
rows inserted in that periode.

mysql> select t, count(*) from t1 group by t order by t;

    +------------+----------+
    | t          | count(*) |
    +------------+----------+
    | 22:40:42.3 |    22528 |
    | 22:40:42.4 |    27104 |
    | 22:40:42.5 |    24256 |
    | 22:40:42.6 |    25344 |
    | 22:40:42.7 |    18912 |
    | 22:40:42.8 |    21408 |
    | 22:40:42.9 |    25088 |
    | 22:40:43.0 |    23328 |
    | 22:40:43.1 |    20928 |
    | 22:40:43.2 |    21408 |
    | 22:40:43.3 |    21312 |
    | 22:40:43.4 |    20640 |
    +------------+----------+

(Refer to bug report for baseline results)

Change-Id: I8b818511f6ea4cc5911663e758b8525f7e433363