Bug#33852756: ESCAPE keyword with non-string values
Bug title is a bit misleading, as the problem may occur
with any type of expression in the ESCAPE clause.

What happens is that the member function eq_cmp_result() on class
Item_func_like tells that with the same non-nullable expression
used both as first and second argument to LIKE, the result is always
TRUE and can be shortcut during optimization. However, this is not
correct when an ESCAPE clause is provided for LIKE, as the escape
character will alter the interpretation of the second argument.

Note also that this optimization is not performed on select list
arguments, thus we see different results when LIKE is used in
select list versus in the WHERE clause.

The fix is to skip this optimization when using an ESCAPE character.

Change-Id: Ibe8da31dc5d5cf4c1114532a22f00bfd52728899