WL#15060: Change group_replication_transaction_size_limit default in 5.7
The option `group_replication_transaction_size_limit` limits the
size of transactions allowed in a Group Replication primary, that
is, if a given transaction exceeds this value it will be rollback.
This option was added in 5.7.19 and 8.0.2-dmr, since 5.7 was a GA
version it was added with a default value of 0 to avoid behaviour
change. On 8.0 the default value is 150000000.

When its value is 0 the memory used for transactions is unbounded,
which can cause out of memory errors.
To prevent these errors we did change the 5.7 default to
150000000, the already default value on 8.0.

RB: 27623