WL#15070 Configuration of NDB replica batch size
* High Level View

The approximate number of bytes in a batch when writing to NDB is
currently controlled by session variable ndb_batch_size. The replica
will use the global value of that variable which means that changing
batch size for the replica also affects the value used by all
sessions.

* Implementation Notes

Adds a new ndb_replica_batch_size variable, letting it control the
ndb_batch_size used for the replica. The new variable has a higher
default value (increased 64-fold to 2MB) to offer a better replication
experience. Overall, the new setting allows the user to fine tune the
performance of replication without affecting the behavior of the
source binary log applier. A set of tests were added to showcase the
behavior.

Change-Id: I36073d118f06dd37d84c7ace0601675ad149477b