Bug#33855163 / Bug#33825393 crashes in ha_ndbinfo::unpack_record()
In NdbInfoScanVirtual, pad string values in buffers to a 4-byte
boundary so that numeric values following string values will be
aligned to a word boundary.

In ha_ndbinfo.cc, create 8- and 16-bit values for enum/set columns
before unpacking them, to assure correct byte order. If any
errors are encountered in unpack_record(), return the code
HA_ERR_INTERNAL_ERROR from rnd_next().

In the test ndbinfo_plans, select the table_event column from an
ndbinfo.events record. For this test to pass on a Sparc machine,
the byte ordering in a SET column must be interpreted correctly.

Change-Id: I3ed41c422c2a76beaae7612e27e83686646c7393