Bug #33874978 GR notification thread can get stuck when new nodes are added to the Cluster
When new nodes are added to the Cluster the Router reconfigures the notification
connections via the x-protocol to each of the Cluster nodes. Due to the logic
error the existing connections are unnecessarily removed and recreated.
That in turn leads to a scenario where the fds of those
connections can be released to the OS while the listening thread is still
using it in the poll.
It can sometimes happen that the poll will spuriously be notified about the
data on the fd that has been reassinged and the attempt to read from that socket
will block until there is something to read from this socket.

This patch fixes the logic - now if the new nodes are being added the connections to
the existing ones are not being recreated.
It also prevent the Router from releasing the fd to the OS while the poll can still
be using this fd.

Change-Id: I90e2dc5ddaa395a8e861abf5681cd8772c5133f7