Bug #33574272 Lob's purge may suffer from nested mtr, which cause FSEG NOT FULL list broken.
Problem:

The record to purge contains multiple blobs.  When the first blob is
freed, the first page of the blob is freed by btr_mtr.  This modifies
the allocation bitmap in the xdes page.  When the next blob is freed,
the blob pages other than the first page is freed by the local mtr.  So
they will also modify the allocation bitmap in the xdes page.  We don't
allow the same page to be modified by the nested mtrs because the redo
logs could become out of order.

The assert catches the mtr conflict problem.

Solution:

The btr_mtr should free the first page of blobs after completing all the
local_mtrs of all blobs.

rb#27714 approved by Yasufumi Kinoshita <yasufumi.kinoshita@oracle.com>