Bug#33060953 - UPGRADE FROM 5.7 TO 8.0 CHANGE EXECTION PLAN                IN PARTITIONED TABLE, EVEN AFTER REVERTING                NEW OPTIMIZER FLAG
In WL#4245, all the NOT IN and NOT EXISTS subqueries are
transformed to antijoins. subqueries are transformed to
AJ-nests and subqueries conditions are transformed into
AJ-nests's join_cond() and added as part of LEFT JOIN.

During JOIN::prune_table_partitions() phase, inner part of
nested outer_joins were not considered for pruning table
partitions. Thus ignoring the AJ-nest which was added as
part of the above tranformation.

The predicates for pruning of table must be placed in the
outer-most join nest, look up the join nest hierarchy for
the outermost condition.

Change-Id: I9b51e535158dadafd8d1549ae647fb3a901f9556