Bug#33784241 compression and replica worker test timeout [noclose]
Pending fix for "Relay log position not advanced with
compressed NDB binlog"

Add test checking that relay log position is advanced on the replica
when using ndb_log_transaction_compression, thus avoiding that the
replication sync will timeout when waiting for same position
as source to be reached. Test is written to be independent on
number of workers configured since currently no workers at all
can be used with standard NDB cluster replication.

Change-Id: I58958242e6948f538df4851517f603986f47ca1e