Bug#33270516: 8.0.22+: crash in KEY::has_records_per_key const
When a derived table having an order by on a field which is
part of the select list gets merged to a query block which has
an aggregate distinct, server fails trying to get index statistics.
It fails because the field for which aggregation needs to be
done is marked hidden resulting in not picking the keypart for that
field as part of the keyparts which need to be used later to get
the index statistics.
When a derived table gets merged, all the fields that are using
the merged derived table's fields will be replaced by
Item_view_refs. For all occurrences of a particular merged derived
table field, there will be as many view references created in the
outer query block, but the underlying field will be the same.
In the failing case, the merged derived table has an order by
field which is part of the select list. So the order by field
is made to point to the select list field during resolving.
When the table gets merged, re-resolving order by which is
now part of the outer query block will add the field as hidden
because its not part of the outer query block's select list.
However the Item_view_ref for the aggregating field is pointing
to the same field. While collecting fields that are part
of aggregation functions, group index skip scan strips the
Item_view_ref and uses only the underlying field. This means
we are using a field which is marked hidden and hidden fields
are skipped while collecting keyparts.

Fix is to not check for the hidden status of items. The original
check for skipping of hidden elements was introduced because
optimizer was looking into the complete list of the fields rather than
just the visible fields. It is indeed correct to look only into
the visible fields because we are trying to convert DISTINCT into
GROUP BY. This is inline with several other places where the
checks are being done only on the visible fields.

Change-Id: Ie834d2dac59b79bc5788b858dee8fe76364deabf