Bug #33856332 InnoDB should let HP transaction timeout while waiting on a lock
InnoDB didn't let a High Priority transaction to timeout when waiting for a lock. It simply ignored the time limit being exceeded and kept waiting.
In case of undetected deadlock (involving commit order violation in group replication appliers) that is more damaging than helping.
It seems it's better to let it timeout and be retried, than to have the whole system being stuck in the deadlock forever.
A proper solution to that is probably to properly KILL QUERY of the deadlock's victim, but implementing this will take some time and investigation.
This simple quick-fix will buy us some time reducing the severity of the issue faced by a customer.

This fix lets an HP transaction to give up waiting if it was interrupted (KILL QUERY) or exceeded the timeout, but only if the blocking transaction is also HP.
In case the blocking transaction is non-HP, it will keep on waiting in the hope that the non-HP will eventually release the locks.
(Note: in case the non-HP indeed releases the locks, and HP gets finally woken up, it's query will still end with an error in case of interruption, but this is exactly as it works in trunk now)

This new rule required some small refactoring in the code which is executed by the woken up transaction's thread to determine what was the reason for wakup.
It's no longer enough to check how much time has passed and if current transaction is HP or not, to determine if the reason for wakeup was a timeout, because now it also depends on the priority of the blocker (who might be already gone, and thus difficult to check).
Therefore instead of redundantly checking the clock and own type, we simply rely on the `trx->error_state` being correctly set by the thread which has woken up the trx.
That is, in trunk, it was the duty of the trx itself to set the `error_state` to `DB_LOCK_WAIT_TIMEOUT` conditionally, but with this patch it is the `lock_wait_timeout_thread`'s duty to set it corrrectly before issuing the wake up signal.
I know of 4 possible reasons for wake up:
1. interruption
2. timeout
3. deadlock (can be "faked" by HP transaction)
4. succesfully obtaining the lock
(5. the transaction itself no longer needs the lock in semi-consistent read, but then it hasn't really started waiting)
I've added asserts which prove that timeout and deadlock are mutually exclusive.

Also, as we've replaced `gettimeofday()` with `std::chrono::steady_clock::now()` some time ago, the fix for Bug#36819 is no longer needed, so I've removed it.

RB: 27683
Change-Id: I36279352499a89acdd1cc1c4233bfad8382fc031