WL#15145 ndb binlog writeset trans dependencies [#2]
Enable the NDB binlog injector to calculate transaction
dependencies for changes written to the binlog.

This is done by extending rpl_injector to populate the "THD::writeset"
array with 64 bit hash values representing the key(s) of each binlogged
row in the current transaction. Those values are then at binlog commit
time compared with the historical writeset to find the oldest transaction
for which there are no change to the same key(s).

The same mechanism is already available for other storage engines since
WL#9556 and intention is that it should now work in an identical fashion
and with same limitations also for transactions binlogged for changes
done in NDB.

The transaction writeset hash value calculations is
enabled by using --ndb-log-transaction-dependency=[ON|OFF],
thus enabling the use of WRITESET dependency tracking mode when
the ndb_binlog thread writes to the binlog.

Expected result is basically that the "last_comitted" value
for each binlogged transaction will be set to the sequence
number of the transaction that previously have modified the
same row(s) or the last serial synchronization point.

Change-Id: I7b50365ce13d26a70eb5e93a3d703c0fe82ba8a4