Bug#33913908 Disable replica worker for NDB replication
The work done for WL#14444 enabled NDB replication to use the replica
worker instead of the single threaded code path which was used earlier.
Due to problems where relay log position is not advanced with compressed
binlog, the usage of replica worker will be disabled for now.

Additionally fix some tests to be independent on wheter or not replica
worker is used or not.

Change-Id: I8818a536b07befd9786755d7a12b73bf778e7ecb