WL#14672: Enable the hypergraph optimizer for UPDATE [8/8, hash join]
Enable use of hash join for UPDATE statements when using the
hypergraph optimizer.

With hash joins, the row IDs used for delayed update cannot be
retrieved from the underlying scan, because the hash join iterator
does not guarantee that the underlying scans are positioned on the
correct row. Instead, the hash join iterator is instructed to store
row IDs of the underlying tables in the join buffer.

The existing execution code assumes that UPDATE uses nested loop joins
only, and enables semi-consistent reads on the outermost table when
the isolation level is read committed or lower. Hash joins are not yet
prepared for doing semi-consistent reads, so the patch disables
semi-consistent reads when the outer table is involved in a hash
join.

Change-Id: I8e712cca62f3e9c3beee5c93ddbd84ebcc61aeb1