WL#15071 Configure NDB replica BLOB write batch size
* High Level View

When writing BLOB values to NDB tables these can be batched into one
roundtrip to the data nodes, within the bounds of a transaction. The
number of bytes that limit this batch is controlled by the variable
"ndb_blob_write_batch_bytes". This variable, however, applies to both
replicas and source nodes. Hence, if there is a requirement to use a
shorter value of "ndb_blob_write_batch_bytes" on the source nodes then
the replicas will use this value and thus need to be configured
separately on their own sessions.
Therefore, replicas benefit from their own configuration option which
this commit implements, using a new variable
"ndb_replica_blob_write_batch_bytes".

* Implementation Notes

- The variable is added as a THD variable, even though its use might
  be over the global variable (THD == null on binary log threads on
  the replica side).
- A test was added to showcase the usage of the new variable.
- Additionally fixes tests broken on WL#15070 due to the added variable.

Change-Id: I33f0d8bc36700d1a70122349a995f057ca0c815a