Bug#33732907: Slow query log logs password in plain text on syntax error
If a statement could not be parsed, it would not be included in
the general log (unless --log-raw was used, in which case the
statement would be logged before parsing).

By contrast, the slow query log would include such a query,
which in theory could expose sensitive credentials.

This patch changes the behaviour so that unparseable queries are
not slow-logged. This aligns it with the default behaviour of
the general log. The slow query log will not observe --log-raw
or align its behaviour with the performance schema's, as it is
not the correct tool to debug malformed queries or hunt down
possible attempts at SQL injection.

Backport from 8.x

Change-Id: I2395bc853b8c66a60ab6e1578dceffbb48de2c05