Fix bug in REPLACE() in UNION truncating result
Item_func_replace::resolve_type() incorrectly calculates char_length,
which propagates to max_length and truncates the result.

Conditions:
UNION query
Longest field (whether physical field, string literal, or the
forumla "char_length * (replace_length - 1ULL)" in the "resolve_type()"
function ) determines the max_size for all values in the result column.

So if a REPLACE() extends string length beyond that size, it will be
truncated.

e.g. by the old formula, using the query below, the char_length for the
REPLACE = 20, and the result string is 19 characters long, so it is not
truncated.

char_length = char_length * (replace_length - 1ULL)
20 = 4 * (6-1)

SELECT REPLACE('aaab','a','123456') as 'L19' UNION SELECT REPEAT('a',12);
+---------------------+
| L19                 |
+---------------------+
| 123456123456123456b |
| aaaaaaaaaaaa        |
+---------------------+
// NOT TRUNCATED

But with the following, the result string is 24 characters long and will
be truncated to 20.
SELECT REPLACE('aaaa','a','123456') as 'L24' UNION SELECT REPEAT('a',12);
+----------------------+
| L24                  |
+----------------------+
| 12345612345612345612 |
| aaaaaaaaaaaa         |
+----------------------+
// 4 CHARS TRUNCATED

And, finally, if we increase the length of the other part of the UNION
to 24, then we will not see any truncation.
SELECT REPLACE('aaaa','a','123456') as 'L24' UNION SELECT REPEAT('a',24);
+--------------------------+
| L24                      |
+--------------------------+
| 123456123456123456123456 |
| aaaaaaaaaaaaaaaaaaaaaaaa |
+--------------------------+
// REPEAT() EXTENDS SIZE to 24, REPLACE() NOT TRUNCATED

*************
**SOLUTIONS**
*************
**1.)** match exact length of output string, using division:
char_length = char_length / search_length * (replace_length - search_length)
        + char_length;

2.) Just remove "- 1ULL" from original formula:
char_length = char_length * replace_length

3.) adjust args[0]->max_char_length somehow