Bug#33893369 static bool PFS_key_pstring::stateless_match(): Assertion `false' failed.
Problem
=======

Before this fix, the query:

  SELECT * FROM
    performance_schema.pfs_example_country FORCE INDEX(name)
    WHERE name='foo1';

failed with an assert in debug builds.

Root cause
==========

Table performance_schema.pfs_example_country
exposes a multi part index:
  UNIQUE KEY(NAME, CONTINENT)

Using the clause FORCE INDEX(name) with WHERE name='foo1'
forces the optimizer to perform a partial key index fetch,
with:
- NAME = 'foo1'
- CONTINENT = N/A, no value provided

In the execution of Country_index_by_name::match(Country_record *record),
this leads to a failed assert, when trying to compare:
- the actual CONTINENT column of the row inspected
- the value from the CONTINENT field in the index key.

The problem is that the code should not get that far,
and not attempt to compare the second key part when not provided.

Fix
===

In country_index_read(), after extracting each key part value,
the fix is to add a call to service
  pfs_plugin_table_v1 -> get_parts_found()
to preserve how many key parts are found in the index key,
instead of assuming a full index key (NAME + CONTINENT) was provided.

This is preserved in Continent_index::m_fields

In Country_index_by_name::match(),
the number of fields seen is checked before making
a comparison for each field value.

Now, to make the code build, the service used to implement
performance schema plugin tables should be "pfs_plugin_table_v1",
which is current and exposes get_parts_found().

The code was using the deprecated and incomplete service
"pfs_plugin_table" instead, which does not expose get_parts_found().

This change upgrades all the implementation of the pfs_example component
to use the proper service, pfs_plugin_table_v1, with associated services
for each column data types.

As a result, many lines were changed.

For example:
  mysql_service_pfs_plugin_table->add_tables(...)
is replaced with:
  pt_srv->add_tables(...)

For columns:
  mysql_service_pfs_plugin_table->get_field_year(...)
is replaced with:
  pc_year_srv->get(...)

The shorter syntax is achieved thanks to the REQUIRES_SERVICE_AS helper:
  REQUIRES_SERVICE_AS(pfs_plugin_table_v1, pt_srv)
  REQUIRES_SERVICE_AS(pfs_plugin_column_year_v1, pc_year_srv)

Change-Id: I1262021e6decd6213c439fa7a9af386074607b1d