WL#14690: Support IF EXISTS and IGNORE UNKNOWN USER optional clauses in REVOKE statement
The aim of this WL is to support optional "IF EXISTS" and "IGNORE UNKNOWN USER"
clause in REVOKE sql statement. "IF EXISTS" clause will convert errors to
warnings which occur when a privilege mentioned in the statement has not been
granted to the specified users/roles. "IGNORE UNKNOWN USER" clause will convert
errors to warnings which occur when the target user/role in the statement does
not exist.

RB#27519