Bug#33944074 Router stuck on connecting to endpoint blocks the shutdown
The Router refuses to shut down (SIGKILL) once there isa connection attempt
to the non-responding endpoint initiated by the user request.
It only shutdown after the 'connection-timeout' configured expires while
it should do it quickly after the shutdown signal.

The problem is that when the Router shuts down the connecting socket is
never canceled. This patch adds canceling a connecting socket.

Change-Id: If6cbab722e15763a73c4c8667e95a3bad6aff1e5