Bug#33773799: recent regression: TABLE::set_keyread(bool): Assertion `file' failed.
Assert failures were seen in debug builds in some circumstances when
executing prepared statements with subqueries that accessed no tables,
and the evaluation of a subquery raised an error.

The reason was that JOIN::destroy() by mistake took a path intended
only for the hypergraph optimizer.

Fixed by guarding the path better to ensure it is not taken when using
the old optimizer.

Change-Id: I8213bb43474fe0e79d00f713ccb247af435684fd