Bug #33965146 PB2: ndbcluster.information_schema_disk_data : result mismatch
The test creates a tablespace and the associated data file id is
checked with INFORMATION_SCHEMA.FILES and ndbinfo.dict_obj_info.

The test retrieves the id from IS. Then uses data file object type
(which is 22) to get the id from ndbifo :
'SELECT id, fq_name FROM ndbinfo WHERE type = 22;'

If the id from ndbinfo is the same as the one from IS, then the resulting
id is replaced by the text "datafile_objectid" in order to remove the
non-determinism of the id.

Problem: When the id=22, '--replace_result' replacs not only the query
results, but also the query text 'type=22' to 'type = datafile_objectid'.

Fix: Read id from IS and ndbinfo and asserts that they agree.

Change-Id: I94e313b75acd18ddafc06a22ad51326c933e202f