Bug#33957087: Fix clang-tidy, cppcheck warnings
The scope of this bug is to fix a selection of clang-tidy and cppcheck
warnings related to the MySQL server code.
In general, the aim is to fix large number of non-controversial, easy to fix issues
(no major code refactoring needed), focusing on server general code.

Post-push fix: correct fix for unsigned variable compared against >= 0.

Change-Id: I0994c416d63401727e8b47051dd3bf615b9723cd