Bug#33848486: Assertion failure in join_optimizer.cc
Inconsistent row estimates were seen in some queries that had a cyclic
query graph, and where one of the join predicates could be pushed down
to a REF access, but the join predicate could not be fully subsumed by
the REF access (because the types of the columns in the predicate
didn't match exactly).

The inconsistencies were caused by ApplyDelayedPredicatesAfterJoin()
applying the selectivity of the join predicate even though the
selectivity had already been applied to the REF access below the join
node.

Fixed by making ApplyDelayedPredicatesAfterJoin() not apply the
selectivity of already applied sargable join predicates.

Change-Id: I5cf8b22303e2b6b5927adde9c062a34bec656b17