Bug#33905399: Missing profiling data for CTE
For certain queries using a common table subexpression (CTE), EXPLAIN
ANALYZE does not give any profiling data (actual time and number of rows)
for the CTE even if the CTE is known to be executed. This happens when:
* The CTE is referenced more than once one the query plan.
* The first (in "EXPLAIN format=tree" output order) reference to the CTE is
  never executed.
* At least one of the subsequent references are executed at least once.

The problem is that the CTE plan is always printed when we encounter
the first reference to the CTE. If that reference is never executed,
the CTE will not be materialized there. And therefore there are
no profiling data to print.

The fix ensures that we print the CTE plan at the first point where it
is executed, which is then the point where it is materialized. The output
will then include profiling data. If the CTE is never executed, we print
the plan at the last reference. Then there are no profiling data.

Change-Id: I81c81c4ec24f3a23a04c8b4398785f68fe9e5d5c