Bug#33975425 routertest_component_routing_splicer fails randomly on pb2
The test "invalid_metada" makes a successfull client connection and then in the
loop keeps making the connections until it fails because of the invalid metadata.
It assumes the failure will always be error 2003 (server unavailable) but depending
on the timing this is sometimes "SSL con aborted" error.

This patch adds SSL error as possible failure cause of the connection.

Change-Id: Ib9a2ff553e9e959be73f9f040b7f541a4b966402