Bug#33975771 routertest_component_routing_strategy.RemoveDestination fails randomly on pb2
The test is making assumptions about the order of round-robin destinations when
doing a new connections. It assumes that the order is kept between the changes
of the destination sets. While this should be a safe assumption most of the
time it fails sometimes on the pb2. What is important is that the round-robin
works well with a stable set of the destinations, and it is not something that
is tested in this test anyway.

This patch changes the expectations so that the checks are not strict more than
necessary by the test logic. The test needs to make sure that given node is or
is not in the round-robin rotation.

Change-Id: I11bc998404331e310aede53f896731368b92b4d1