Bug#33975782 routertest_component_http_server.multiple_backends_works fails randomly on pb2
The test launches the Router with http configuration that uses password file from
the disk. Only after that it creates that file hoping it will be quick enough
to do that before the running Router needs it. It's a clear race condition.

This patch changes the order: creates the password file before the Router is
launched.

Change-Id: Ifec4d4019c90e96cb36ca6c1bdf24f8f928419a2