Bug#33976040 routertest_component_rest_metadata_cache.metadata_status fails randomly on pb2
The test is using sleep(0.5s) assuming it will be long enough for the Router to
do another metadata refresh round. That's too optimistic assumption for pb2 env.

This patch changes the synchronisation in the test replacing sleep with proper
checking in the mock_server that the metadata refresh really happened.

Change-Id: I412fe469e2e3e2b5860b4317dd8bec41fdfdade7