Bug#33725503: A bug in sql_derived.cc caues mysql-server failure
The problem here is similar to the one in bug#32962511. As in the
mentioned bug, there is a recursive CTE and a removal of a query
expression after const predicate elimination. It would thus be
expected that when the reference count of TABLE objects for the
CTE temporary table reaches zero, it should be possible to
recreate the table again.

However, in this case one of the TABLE references is not recorded as
attached to the CTE, thus in the test

  if (m_common_table_expr && m_common_table_expr->tmp_tables.size() > 0)

in TABLE_LIST::setup_materialized_derived_tmp_table(), tmp_tables.size()
is greater than zero and we will attempt to create a cloned tmp table
instance, even though the temporary table does not exist.

The fix is to record a cloned temporary table as belonging to a CTE,
so that we know better when to decrement the reference count.

Change-Id: I68f04504d49ddf65a76d92502351446f967960d0