Bug#33834514: 'RAND(0)' gives inconsistent results with hypergraph
The hypergraph optimizer used to rematerialize uncorrelated derived
tables for each row in the outer table if the derived table was
non-deterministic. The old optimizer materializes it once. The results
therefore differed between the two optimizers.

Fixed by changing the hypergraph optimizer to rematerialize the
derived table only if it is correlated (which in practice means that
it's a table function or a lateral derived table).

Change-Id: I030ec1c3906753862537026c24851733e98681a2