Bug#33952115: Semijoin may give wrong result Bug#33957233: Incorrect inner hash join when using materialization
These two bugs have the same cause.
The symptoms of the problem is incorrect results from execution
of semi-join with materialization. It may happen when there is
an equality in the WHERE clause of the subquery. In some cases,
such as when one side of the equality is an IN or NOT IN subquery,
the equality is neither pushed down to the materialized subquery,
nor evaluated as part of the semi-join. This happens because the
equality remains in the semi-join evaluation, and its field references
are only partially substituted with field references from the
materialized subquery. The problem with this is that the equality gets
a table map that consists of tables from both the outer tables and the
subquery tables, and the optimizer is unable to push such condition
into any table condition.

The solution to the problem is however to realize that such conditions
can be evaluated when materializing the subquery. If the condition's
table map only contains tables from the subquery, no substitution
should take place, and the condition will be properly pushed down to
the subquery materialization.

Another small fix was also included: For a triggered condition,
it was forgotten to add triggered function tables in
fix_after_pullout().

Change-Id: I4fc1a2d0a280e3675ea1824ef88f078ee667b23e