Bug#33986557 routertest_integration_routing_reuse times out on Windows.
The test is launching the mysqld process once and the reuses it for each testcase.
It does not do that for the mysqlrouter process which means that it is being
spawned and killed for each testcase (1000+ times). That takes too long for the pb2,
and if fails the tests because of the timeout.

This patch changes the test so that a single Router process is reused for
each testcase, similar way the mysqld process is reused.

Change-Id: If61a783af60e7fbef39d5b6b0ec8eba0fd947b2a