Bug#33550360: Inconsistent row estimates in the hypergraph optimizer [noclose]
Equivalent join predicates created from multiple equalities could in
some cases be added twice to an edge in the hypergraph, both as a=b
and as b=a. The selectivity of this predicate would be counted twice,
which would cause inconsistent row estimates.

Fixed by changing AlreadyExistsOnJoin() to also look for equalities
with the operands in the opposite order.

Also removed some duplicate code by using existing helper functions.

Change-Id: I561b9bf7ff558d526860194d0f005ec29b08fa21