Bug #33972317 Server 8.0 MSI is missing INFO_BIN and INFO_SRC
In pushbuild we have *two* copies of the file INFO_SRC

 - ${CMAKE_SOURCE_DIR}/Docs/INFO_SRC
   This is created when the source tarball is created.
   It will be installed with this cmake code:
   INSTALL(DIRECTORY Docs/ DESTINATION ${INSTALL_DOCDIR}

 - ${CMAKE_BINARY_DIR}/Docs/INFO_SRC
   This is created during cmake configure/build.
   It will be installed with this cmake code:
   INSTALL(FILES
     ${CMAKE_BINARY_DIR}/Docs/INFO_SRC
     ${CMAKE_BINARY_DIR}/Docs/INFO_BIN
     DESTINATION ${INSTALL_DOCDIR} COMPONENT Info

The result is that we have two entries with the same name,
and creation of MSI or ZIP builds on Windows will fail.

${CMAKE_BINARY_DIR}/Docs/INFO_SRC is removed by build scripts in
pushbuild, thus avoiding the file duplication.

This patch removes ${CMAKE_SOURCE_DIR}/Docs/INFO_SRC
before building the MSI target.

With this removal in place, we can add the 'Info' COMPONENT
to MSI builds on Windows, which is what this bug# was really about.

Change-Id: I0adc8d9f9f0bf6e419150d237fc0ebd2826813b6